requirements.txt
# Required Packages
speedtest-cli

# Standard Libraries Used (no need to install)
# tkinter
# logging
# datetime